+++ example/lib/main.dart
@@ -8,6 +8,7 @@ void main() {
   Telemetrydecksdk.start(
     const TelemetryManagerConfiguration(
         appID: "22385F1C-3699-4F04-9D63-24CC0B2E62D8",
+        namespace: "my-namespace",
         defaultParameterPrefix: "DemoApp.Parameter.",
         defaultSignalPrefix: "DemoApp.Signal.",
         defaultParameters: {"ParameterName": "ParameterValue"},